move auto loader to separate file
